package renamed, useless lines deleted :P
