+++ MyCalculator/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.totha.mycalculator">
+    package="com.calculator">
 
     <application
         android:allowBackup="true"
@@ -8,16 +8,14 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
-            <!--android:screenOrientation="portrait"
-            android:configChanges="keyboardHidden|orientation|screenSize"-->
+
+        <activity android:name="titans.calculator.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ShowResultsActivity"></activity>
+        <activity android:name="titans.calculator.ShowResultsActivity"/>
     </application>
-
 </manifest>
