+++ MyCalculator/app/src/main/java/titans/calculator/PolishFormAlgorithm.java
@@ -1,4 +1,4 @@
-package com.example.totha.mycalculator;
+package titans.calculator;
 
 import android.support.annotation.NonNull;
