+++ MyCalculator/app/src/main/java/titans/calculator/ShowResultsActivity.java
@@ -1,4 +1,4 @@
-package com.example.totha.mycalculator;
+package titans.calculator;
 
 import android.content.Intent;
 import android.graphics.Color;
@@ -6,6 +6,9 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
+import com.calculator.R;
+
 import java.util.HashMap;
 import java.util.Map;
