+++ MyCalculator/app/src/main/res/layout/activity_main.xml
@@ -1,13 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/activity_vertical_margin"
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.totha.mycalculator.MainActivity"
     android:orientation="vertical"
     android:background="#444444"
     android:weightSum="1">
@@ -64,7 +62,7 @@
                     android:layout_weight="1"
                     android:text="@string/seven"
                     android:onClick="onClick"
-                    style="@style/MyNumberButtonTheme"/>
+                    android:theme="@style/MyNumberButtonTheme"/>
                 <Button
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
