+++ MyCalculator/app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">MyCalculator</string>
+    <string name="app_name">Titans Calculator</string>
     <string name="clear">Clear</string>
     <string name="delete">Delete</string>
     <string name="one">1</string>
